refactor(custom-css): move migration helpers to shared utils
