+++ src/renderer/pages/settings/AppearanceSettings/AppearanceSettings.tsx
@@ -36,9 +36,9 @@ import { toast } from '@renderer/services/toast'
 import { formatErrorMessage } from '@renderer/utils/error'
 import { isLinux, isMac } from '@renderer/utils/platform'
 import { cn } from '@renderer/utils/style'
-import { hasV1CustomCssMarker, removeV1CustomCssMarker } from '@shared/data/preference/customCss'
 import type { LanguageVarious, MenuPresentationMode } from '@shared/data/preference/preferenceTypes'
 import { ThemeMode } from '@shared/data/preference/preferenceTypes'
+import { hasV1CustomCssMarker, removeV1CustomCssMarker } from '@shared/utils/customCssMigration'
 import { defaultLanguage } from '@shared/utils/languages'
 import { AlertTriangle, Download, Minus, Monitor, Moon, Plus, Sun } from 'lucide-react'
 import type React from 'react'
