+++ src/renderer/pages/settings/AppearanceSettings/__tests__/AppearanceSettings.test.tsx
@@ -1,7 +1,7 @@
 import { popup } from '@renderer/services/popup'
 import { toast } from '@renderer/services/toast'
-import { V1_CUSTOM_CSS_MARKER } from '@shared/data/preference/customCss'
 import type { MenuPresentationMode } from '@shared/data/preference/preferenceTypes'
+import { V1_CUSTOM_CSS_MARKER } from '@shared/utils/customCssMigration'
 import { MockUsePreferenceUtils } from '@test-mocks/renderer/usePreference'
 import { fireEvent, render, screen, waitFor } from '@testing-library/react'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
